Append user approval files to backup directory.
Instead of creating new directories in the backup directory which hold
the approved.users and mid_approved.users files as they were at the time
the script was run, just append the files to "approved.users" and
"mid_approved.users" in the backup directory.

This keeps all backups in two files which can easily be integrated with
logrotate in order to automatically delete old files when they're no
longer needed. This is desirable because these files contain encrypted
user passwords, which only need to be stored until the request is
processed, at which time the password is hashed and the encrypted copy
should be securely discarded.

The files are simply appended, which means that it's possible that
duplicate entries will appear in the backup files if, for example, an
account is not processed. This is not really a concern since the backup
files are only for human use. However, a comment precedes the files,
making it easy to see the state of a file prior to a given run.