filter_accounts: fix indexing of empty list
4cd57395 cleaned up some code, but accidentally starting using results
before checking that it wasn't an empty list.

It's been broken for months but we didn't actually notice because nobody
deployed the code. I noticed today because there was a bug in the
similarity function which Dara had already fixed. When I deployed the
latest version, it fixed the similarity function, but then this popped
up.